Apply patch from Sven Fabricius to use PAC
Hi Sean,

I've a patch for your GitHub Repo. I worked on implementing the PAC (privilege attribute certificate) which is supplied via the Kerberos token data.
It is very useful; the web application can use this for validating group memberships without asking a domain controller.

The PAC is parsed to a XML file and stored to the cache directory. The filename compares to the username + .xml .

I thought about to return the filename to the nginx instance as variable, so it can be passed to a cgi instance, maybe as .
But I'm not so deep in nginx to implement this.

I've added three new options:
auth_gss_pac (on/off) : parse the PAC or not
auth_gss_pac_cache_dir (path + trailing slash) : the path where to cache the PAC XML (default: /tmp/)
auth_gss_pac_cache_time (number) : time in seconds, were the cache file will not be updated, to save performance (default: 1 hour - 3600 seconds)

I'm not firm with GitHub, so I send you the patch directly.

I hope you enjoy it. Feel free to improve the source.

Regards Sven