Reduce log level for debug messages
Debians nginx is built using --with-debug, meaning that these messages will be
printed unconditionally if they're at level "INFO". Also, there are some places
that use NGX_LOG_INFO and some that use NGX_LOG_DEBUG in the same
"if (NGX_DEBUG)" block, so this makes the code more consistent.